[syntax] Remove v2-b and rename v2-a to v2
The syntax package has remnants of a multipart ASTv2 process that never
came to fruition.

This commit removes those remnants.
