--- packages/@glimmer/syntax/lib/v2-b/README.md
@@ -1,1 +1,0 @@
-The second step in ASTv2 normalization converts local variable lookups into allocated symbols.
