+++ packages/@glimmer/syntax/lib/v2/loose-resolution.ts
