+++ packages/@glimmer/syntax/lib/v2/normalize.ts
