+++ packages/@glimmer/syntax/lib/v2/objects/NAMING.md
