+++ packages/@glimmer/syntax/lib/v2/objects/attr-block.ts
