+++ packages/@glimmer/syntax/lib/v2/objects/base.ts
