+++ packages/@glimmer/syntax/lib/v2/objects/content.ts
