+++ packages/@glimmer/syntax/lib/v2/objects/ebnf/wfc.md
