+++ packages/@glimmer/syntax/lib/v2/objects/internal-node.ts
