+++ packages/@glimmer/syntax/lib/v2/objects/node.ts
