+++ packages/@glimmer/syntax/lib/v2/objects/resolution.ts
